makes max window size std::f32::MAX

+++ src/platform_impl/macos/window.rs
@@ -509,8 +509,8 @@ impl UnownedWindow {
     pub fn set_max_inner_size(&self, dimensions: Option<Size>) {
         unsafe {
             let dimensions = dimensions.unwrap_or(Logical(LogicalSize {
-                width: std::f64::MAX,
-                height: std::f64::MAX,
+                width: std::f32::MAX as f64,
+                height: std::f32::MAX as f64,
             }));
             let dpi_factor = self.hidpi_factor();
             set_max_inner_size(*self.ns_window, dimensions.to_logical(dpi_factor));
